the end... continue...

+++ app/src/main/java/es/usj/mastertsa/pluengo/marvelcollection/core/IDao.kt
@@ -3,9 +3,9 @@ package es.usj.mastertsa.pluengo.marvelcollection.core
 interface IDao<T> {
 
     fun insert(element: T) : T
-    fun update(element: T) : Long?
-    fun delete(id: Long) : T?
+    fun update(element: T) : Int?
+    fun delete(id: Int) : T?
     fun list() : List<T>
-    fun findById(id: Long) : T?
+    fun findById(id: Int) : T?
 
 }
