+++ app/src/main/java/es/usj/mastertsa/pluengo/marvelcollection/core/dao/room/ComicRoomDao.kt
@@ -1,8 +1,8 @@
 package es.usj.mastertsa.pluengo.marvelcollection.core.dao.room
 
-import es.usj.mastertsa.pluengo.marvelcollection.MainActivity
+import es.usj.mastertsa.pluengo.marvelcollection.controller.comicList.MainActivity
 import es.usj.mastertsa.pluengo.marvelcollection.core.IDao
-import es.usj.mastertsa.pluengo.marvelcollection.core.dao.room.model.ComicRoom
+import es.usj.mastertsa.pluengo.marvelcollection.model.ComicRoom
 import es.usj.mastertsa.pluengo.marvelcollection.core.dao.room.utils.ComicIDao
 import es.usj.mastertsa.pluengo.marvelcollection.core.dao.room.utils.MarvelRoomDatabase
 
@@ -21,12 +21,12 @@ class ComicRoomDao() :
         return element
     }
 
-    override fun update(element: ComicRoom): Long? {
+    override fun update(element: ComicRoom): Int? {
         comicIDao!!.update(element)
         return 1
     }
 
-    override fun delete(id: Long): ComicRoom? {
+    override fun delete(id: Int): ComicRoom? {
         val event = findById(id)
         comicIDao!!.delete(event)
         return event
@@ -36,7 +36,7 @@ class ComicRoomDao() :
         return comicIDao!!.getAllComics()
     }
 
-    override fun findById(id: Long): ComicRoom? {
+    override fun findById(id: Int): ComicRoom? {
         return comicIDao!!.findById(id)
     }
